Modied type data for clicks var

+++ url_shortener-backend/src/main/java/com/alxmr/url_shortenerbackend/dto/URLSDto.java
@@ -10,5 +10,5 @@ public class URLSDto {
     private Long id;
     private String originalURL;
     private String shortURL;
-    private Short clicks;
+    private Integer clicks;
 }
